@@ -1,4 +1,4 @@
 from pkgutil import iter_modules
 
 
-EXTENSIONS = [module.name for module in iter_modules(__path__, f"{__package__}.")]
+EXTENSIONS = [mod_info.name for mod_info in iter_modules(__path__, f"{__package__}.")]
