@@ -1,6 +1,7 @@
 import asyncio
 import functools
 import re
+from collections.abc import Callable
 from typing import Any
 
 import aiohttp
@@ -45,13 +46,10 @@ async def on_bad_9gag_link(bot: core.Beira, message: discord.Message) -> None:
                 f"Reposted from {message.author.mention} ({message.author.name} - {message.author.id}):\n\n"
                 f"{new_links}"
             )
-            await message.reply(
-                content,
-                allowed_mentions=discord.AllowedMentions(users=False, replied_user=False),
-            )
+            await message.reply(content, allowed_mentions=discord.AllowedMentions(users=False, replied_user=False))
 
 
-def make_listeners(bot: core.Beira) -> tuple[tuple[str, functools.partial[Any]], ...]:
+def make_listeners(bot: core.Beira) -> tuple[tuple[str, Callable[..., Any]], ...]:
     """Connects listeners to bot."""
 
     # Adjust the arguments for the listeners and provide corresponding event name.
