@@ -1,15 +1,9 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
+import core
 
 from .snowball import SnowballCog
 
 
-if TYPE_CHECKING:
-    from core import Beira
-
-
-async def setup(bot: Beira) -> None:
+async def setup(bot: core.Beira) -> None:
     """Connects cog to bot."""
 
     await bot.add_cog(SnowballCog(bot))
